Move settings back to fix interaction with Layer backend
