+++ layer/training/runtime/model_trainer.py
@@ -9,8 +9,8 @@ from uuid import UUID
 
 from layerapi.api.entity.model_train_status_pb2 import ModelTrainStatus
 
-from layer import Context
 from layer.clients.layer import LayerClient
+from layer.context import Context
 from layer.contracts.assertions import Assertion
 from layer.exceptions.exception_handler import exception_handler
 from layer.exceptions.exceptions import LayerFailedAssertionsException
