+++ test/e2e/test_function_dataset_run.py
@@ -2,12 +2,18 @@ import pandas as pd
 import pytest
 
 import layer
-from layer import Dataset, global_context
 from layer.clients.layer import LayerClient
+from layer.contracts.datasets import Dataset
 from layer.contracts.fabrics import Fabric
 from layer.contracts.projects import Project
 from layer.decorators import dataset, pip_requirements
 from layer.exceptions.exceptions import LayerClientException
+from layer.global_context import (
+    current_project_name,
+    default_fabric,
+    get_active_context,
+    get_pip_packages,
+)
 from test.e2e.assertion_utils import E2ETestAsserter
 from test.e2e.conftest import _cleanup_project
 
@@ -84,10 +90,10 @@ def test_multiple_inits_switch_context(
     prepare_data()
 
     # then
-    assert global_context.current_project_name() == second_project_name
-    assert global_context.default_fabric() == Fabric.F_XSMALL
-    assert global_context.get_pip_packages() == ["tensorflow==2.3.2"]
-    assert global_context.get_active_context() is None
+    assert current_project_name() == second_project_name
+    assert default_fabric() == Fabric.F_XSMALL
+    assert get_pip_packages() == ["tensorflow==2.3.2"]
+    assert get_active_context() is None
 
     # and when
     layer.init(initialized_project.name)
@@ -95,7 +101,7 @@ def test_multiple_inits_switch_context(
     # then
     with pytest.raises(LayerClientException, match=r"Dataset not found.*"):
         layer.get_dataset(dataset_name).to_pandas()
-    assert global_context.get_active_context() is None
+    assert get_active_context() is None
 
     assert (
         len(
@@ -105,9 +111,9 @@ def test_multiple_inits_switch_context(
         )
         == 3
     )
-    assert global_context.current_project_name() == initialized_project.name
-    assert global_context.default_fabric() is None
-    assert global_context.get_pip_packages() is None
-    assert global_context.get_active_context() is None
+    assert current_project_name() == initialized_project.name
+    assert default_fabric() is None
+    assert get_pip_packages() is None
+    assert get_active_context() is None
 
     _cleanup_project(client, project)
